move user adding to its own component
